Better UI elements and equip with Bootstrap 4
Changes to HTML and JS, not PHP.

Aditionally loading font-awesome for some cheeky icons.

Turning `span` objects into `button` objects, adding the `disabled` attribute and a font-awesome spinning gear (`fa fa-cog fa-spin`).

Adding a new argument to poll (`timeout`) and setting it to 10000. Setting that to 0 if status is `waiting1` will result in the progress to be displayed nearly immediately.